requirements.txt
attrs==22.2.0
colorama==0.4.6
daiquiri==3.2.1
DateTime==5.0
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
py==1.11.0
pytest==7.2.2
pytest-html==3.2.0
pytest-metadata==2.0.4
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
tomli==2.0.1
zope.interface==5.5.2
